% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automate_sample_pooling.R
\name{automate_sample_pooling}
\alias{automate_sample_pooling}
\title{Automate Sample Pooling}
\usage{
automate_sample_pooling(Quant, Date_Project, Max_Volume = 8, Min_Volume = 2)
}
\arguments{
\item{Quant}{A data frame with columns: Sample Name, Quantification, Position, Plate, Group (optional).}

\item{Date_Project}{A string representing the name of the project (used for output directories and file names).}

\item{Max_Volume}{The maximum volume allowed for each sample in the pool (default = 8 μL).}

\item{Min_Volume}{The minimum volume allowed for each sample in the pool (default = 2 μL).}
}
\description{
Automates sample pooling, creates plots, and generates Epmo robot output. This function wraps around the internal pool creation, plotting, and robot output functions.
}
