% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_epmo_input.R
\name{create_epmo_input}
\alias{create_epmo_input}
\title{Create Robot Input for Epmo}
\usage{
create_epmo_input(create_pools_df, Theoretical_Pools, Date_Project)
}
\arguments{
\item{create_pools_df}{A data frame from the function create_pools() specifying pool and pooling information,}

\item{Theoretical_Pools}{A data frame summarizing the final volume and theoretical concentration of each pool.}

\item{Date_Project}{A string representing the name of the project.}
}
\value{
No return value; this function saves a dataframe file for Epmo.
}
\description{
Generates the CSV file required by the Epmo robot for automated sample pooling.
}
