% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pools.R
\name{create_pools}
\alias{create_pools}
\title{Create Pools}
\usage{
create_pools(Quant, Max_Volume = 8, Min_Volume = 2)
}
\arguments{
\item{Quant}{A data frame with columns "Sample Name", "Quantification", "Position", "Plate", and "Group". "Group" is optional and allows separate pooling within each group.}

\item{Max_Volume}{The maximum volume allowed for a sample in a pool (default = 8 μL).}

\item{Min_Volume}{The minimum volume allowed for a sample in a pool (default = 2 μL).}
}
\value{
A data frame containing the pooled samples, including pool volumes and pool numbers.
}
\description{
This function takes a data frame of sample quantification data and pools the samples based on their quantification values. It ensures that pool volumes are between the specified minimum and maximum values.
}
