% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plate_maps.R
\name{generate_plate_maps}
\alias{generate_plate_maps}
\title{Generate Pooling Plate Maps}
\usage{
generate_plate_maps(create_pools_df, Date_Project)
}
\arguments{
\item{create_pools_df}{A data frame from the function create_pools() specifying pool and pooling information,}

\item{Date_Project}{A string representing the name of the project (used to name the output directory and files).}
}
\value{
No return value; this function saves plot files as PNG images.
}
\description{
This function creates and saves pooling plots for each plate. It generates a general map for each plate and pool/plate-specific maps.
}
