% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_automate_sample_pooling_input.R
\name{validate_automate_sample_pooling_input}
\alias{validate_automate_sample_pooling_input}
\title{Validate Input for Automate Sample Pooling Function}
\usage{
validate_automate_sample_pooling_input(
  Quant,
  Date_Project,
  Max_Volume,
  Min_Volume
)
}
\arguments{
\item{Quant}{A data frame containing sample information. It must include the columns:
"Sample Name", "Quantification", "Position", "Plate". The "Group" column is optional.}

\item{Date_Project}{A string representing the project name, used for output file naming.}

\item{Max_Volume}{A numeric value indicating the maximum volume allowed in the pools. Must be greater than 0.}

\item{Min_Volume}{A numeric value indicating the minimum volume allowed in the pools. Must be greater than 0 and less than Max_Volume.}
}
\value{
A validated data frame, possibly with a default "Group" column added.
}
\description{
This function validates the input parameters for the `automate_sample_pooling` function.
It checks that the provided data frame has the required columns, ensures numeric constraints on volume parameters,
and verifies that the project date is a valid non-empty string.
}
\keyword{internal}
