% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automate_sample_pooling.R
\name{automate_sample_pooling}
\alias{automate_sample_pooling}
\title{Automate Sample Pooling}
\usage{
automate_sample_pooling(Quant, Date_Project, Max_Volume = 8, Min_Volume = 2)
}
\arguments{
\item{Quant}{A csv file with four columns named "Sample Name", "Quantification" "Position", and "Plate". Sample name can be anything. Quantification should be sample quantification value from quant. Position should be the position within the 96 well plate (A1-H12). Plate should identify which plate the sample is on ("Plate 1"-"Plate N").}

\item{Date_Project}{A character string to specify the name of the pooling run. This will create a new directory with that name and name the files accordingly.}

\item{Max_Volume}{A number specifying the max volume of a sample to be included in a pool (Default = 8ul).}

\item{Min_Volume}{A number specifying the minimim volume of a sample to be included in the pool (Default = 2ul).}
}
\description{
This function is meant to automate sample pooling for sequencing. The process will produce several files including pooling maps, theoretical concentrations, and a script that can be added to the Epmo for completely automated pooling.
}
\examples{
# basic usage of automate_sample_pooling
#For example Usage use:
#For real usage use: Quant_DF <- read.csv(PATH TO CSV FILE)
Quant_DF <- tgen.north.lab.sequencing.tools::Example_Quant_Data
automate_sample_pooling(Quant = read.csv(Quant = Quant_DF, Date_Project = "20240117_Testing", Max_Volume = 10, Min_Volume = 1.6)
}
